+++ takeout.go
@@ -19,6 +19,6 @@ package takeout
 
 const (
 	AppName = "Takeout"
-	Version = "0.14.4"
+	Version = "0.14.5"
 	Contact = "takeoutfm.com"
 )